### 2020-05-28 d078b31
added doi's for paper.bib
