+++ paper/paper.bib
@@ -5,7 +5,7 @@
   volume={1},
   number={1},
   year={2018},
-  doi={https://doi.org/10.5334/joc.18}
+  doi={https://doi.org/10.5334/joc.18},
   publisher={Ubiquity Press}
 }
 
