+++ paper/paper.bib
@@ -5,7 +5,7 @@
   volume={1},
   number={1},
   year={2018},
-  doi={https://doi.org/10.5334/joc.18},
+  doi={10.5334/joc.18},
   publisher={Ubiquity Press}
 }
 
@@ -25,7 +25,7 @@
   author={Hess, Eckhard H and Polt, James M},
   journal={Science},
   volume={132},
-  doi={https://doi.org/10.1126/science.132.3423.349},
+  doi={10.1126/science.132.3423.349},
   number={3423},
   pages={349--350},
   year={1960},
@@ -61,7 +61,7 @@
   author={Aston-Jones, Gary and Cohen, Jonathan D},
   journal={Annu. Rev. Neurosci.},
   volume={28},
-  doi={https://doi.org/10.1146/annurev.neuro.28.061604.135709},
+  doi={10.1146/annurev.neuro.28.061604.135709},
   pages={403--450},
   year={2005},
   publisher={Annual Reviews}
@@ -97,7 +97,7 @@
   number={2},
   pages={252--269},
   year={2010},
-  doi={https://doi.org/10.3758/cabn.10.2.252},
+  doi={10.3758/cabn.10.2.252},
   publisher={Springer}
 }
 
