### 2020-06-12 #2269 @labarba
👋  @oliviaguest — Could you have a look at this submission and let us know if you can handle it as editor? I realize you are handling four other submissions, but perhaps you can accept this on a 'best effort' basis (we have more than 30 submissions in the backlog).

Title — pypillometry: A Python package for pupillometric analyses
[ Article proof](https://github.com/openjournals/joss-papers/blob/joss.02269/joss.02269/10.21105.joss.02269.pdf)
