### 2020-06-12 #2269 @oliviaguest
Oh, wow — very similar to: https://joss.theoj.org/papers/65fdd6bc59eb558543282e93140eb31e but in Python! Fun!
