### 2020-06-12 #2269 @danielskatz
👋 @oliviaguest - just as a reminder, you can make it easier for the AEiC: if you are willing to edit a submission, issue ```@whedon assign me as editor```, and remove the waitlisted label if there is one.
