### 2020-06-15 #2269 @ihrke
Thanks @oliviaguest for handling the submission! Surprised and glad to see @samhforbes 's PupillometryR - looks like there are finally some good tools emerging for pupillometry! Looking forward to the review process!
