### 2020-06-15 #2269 @ihrke
Ignace Hooge perhaps? He is an expert in eyetracking/pupillometry but I don't know how good his Python is: https://www.uu.nl/staff/ihooge/Contact
