### 2020-06-16 #2269 @ihrke
I looked at the list of JOSS reviewers and found the following people that I think competent to review this submission (in - very rough - order of perceived goodness-of-fit based on their description):

szorowi1
sappelhoff
marcoapintoo
agricolab
lukassnoek
jsheunis
faroit
SebastianoF

let me know if you need more names...
