### 2020-06-16 #2348 @whedon
[REVIEW]: pypillometry: A Python package for pupillometric analyses
**Submitting author:** @ihrke (<a href="http://orcid.org/0000-0003-0205-7353">Matthias Mittner</a>)
**Repository:** <a href="https://github.com/ihrke/pypillometry" target ="_blank">https://github.com/ihrke/pypillometry</a>
**Version:** v1.0.0
