### 2020-06-16 #2348 @oliviaguest
Hi all! 👋  Thank you so much, @samhforbes, @szorowi1, for accepting to review this software package and paper by @ihrke! 😊

Any questions, feedback on the paper, etc., please post here. Any *very* code-specific questions, suggestions, etc., please use the issues in the code repo and link to them from this thread so we can all keep track of them. 🌸

Thank you again, Sams!
