### 2020-06-24 ihrke/pypillometry#1 @samhforbes
Modelling tonic and phasic components
Hi @ihrke,
I've been having some fun playing around with the options you have here.
In the paper you claim a strength of this package is modelling tonic and phasic components of the signal, but (for me at least), the notebook section for that (modelling data) just comes up with TO DO:

I feel like there's scope for some things to be a work in progress, but it might be helpful for complete documentation to be able to see some of these options, but also to be able to test those functions work the way they should
