### 2020-06-24 ihrke/pypillometry#1 @ihrke
That's a fair point. In fact, I am currently writing a methodological paper on a new baseline and response-estimation technique and this technique is already implemented in the package (but not very well documented). I kind of wanted to wait until the paper is finished before officially promoting these functions in the package. But I can see the point that a package user would want to get some kind of idea what this module is about.

Would you be ok if I added some documentation to the pupil-modelling section in the docs while adding a warning that this is an experimental and yet untested part of the package?
