### 2020-06-24 ihrke/pypillometry#1 @samhforbes
Yeah I can fully understand wanting to wait. But I do think especially as this is functionality which is unique to your package (and therefore super cool!), it would really strengthen it, even if only experimental features or a basic example with detailed analyses to come.
