### 2020-06-24 ihrke/pypillometry#2 @samhforbes
Community guidelines
@ihrke I can't currently see a clear community guidelines statement on the documentation website or in the README. Remember some of your users may not be familiar with github etiquette, so if you can state on either or both how to contact you, how to report bugs etc, that would be welcome.
