### 2020-06-24 ihrke/pypillometry#3 @samhforbes
References for JOSS paper
@ihrke The paper looks good, but there are a few references missing DOIs (Joshi et al., Kluyver et al., Laeng et al., and Mittner et al.).
I'd also be grateful if you wouldn't mind updating the PupillometryR reference to the JOSS paper (Forbes, S. H. (2020). PupillometryR: An R package for preparing and analysing pupillometry data. Journal of Open Source Software, 5(50), 2285. https://doi.org/10.21105/joss.02285)
