### 2020-06-24 ihrke/pypillometry#4 @samhforbes
Installation instructions for PyStan
Hi @ihrke,
This is more of a suggestion for you to look at, but getting PyStan going can be an issue for users especially if they move outside of jupyter. Given they have a fair amount of documentation on these issues, perhaps it's worth linking users across or highlighting some of the main troubleshooting tips (needing compilers in windows for example).
