### 2020-06-24 #2348 @samhforbes
Installation instructions: Is there a clearly-stated list of dependencies? Link to more in depth installation instructions for PySTan:
https://github.com/ihrke/pypillometry/issues/4

 Example usage and functionality documentation
ideally would have documentation on modelling data. https://github.com/ihrke/pypillometry/issues/1

https://github.com/ihrke/pypillometry/issues/2

 References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
DOIs need to be included for a number of papers. See https://github.com/ihrke/pypillometry/issues/3
