### 2020-06-25 ed91ccb
minor bug in plotting FakePupilData
