### 2020-06-25 97850e8
Fixed references for JOSS paper
- Fixed missing DOIs in references for JOSS paper
- Fixed pupillometryR ref

See also: #3
