+++ paper/paper.bib
@@ -15,6 +15,7 @@
   journal={Perspectives on psychological science},
   volume={7},
   number={1},
+  doi={10.1177/1745691611427305},
   pages={18--27},
   year={2012},
   publisher={Sage Publications Sage CA: Los Angeles, CA}
@@ -46,15 +47,20 @@
         year = {2020},
         file = {Springer Full Text PDF:/Users/mmi041/Zotero/storage/Z3L9FUWR/Geller et al. - 2020 - GazeR A Package for Processing Gaze Position and .pdf:application/pdf}
 }
-
-@Manual{forbes2020,
-    title = {PupillometryR: A Unified Pipeline for Pupillometry Data},
-    author = {Samuel Forbes},
-    year = {2020},
-    note = {R package version 0.0.2},
-    url = {https://CRAN.R-project.org/package=PupillometryR},
-}
   
+@article{forbes2020,
+  doi = {10.21105/joss.02285},
+  url = {https://doi.org/10.21105/joss.02285},
+  year = {2020},
+  publisher = {The Open Journal},
+  volume = {5},
+  number = {50},
+  pages = {2285},
+  author = {Samuel H. Forbes},
+  title = {PupillometryR: An R package for preparing and analysing pupillometry data},
+  journal = {Journal of Open Source Software}
+}
+
   
 @article{aston2005integrative,
   title={An integrative theory of locus coeruleus-norepinephrine function: adaptive gain and optimal performance},
@@ -71,6 +77,7 @@
   title={Relationships between pupil diameter and neuronal activity in the locus coeruleus, colliculi, and cingulate cortex},
   author={Joshi, Siddhartha and Li, Yin and Kalwani, Rishi M and Gold, Joshua I},
   journal={Neuron},
+  doi={10.1016/j.neuron.2015.11.028},
   volume={89},
   number={1},
   pages={221--234},
@@ -85,6 +92,7 @@
   volume={20},
   number={8},
   pages={570--578},
+  doi={10.1016/j.tics.2016.06.004},
   year={2016},
   publisher={Elsevier}
 }
@@ -105,6 +113,7 @@
   title={Jupyter Notebooks-a publishing format for reproducible computational workflows.},
   author={Kluyver, Thomas and Ragan-Kelley, Benjamin and P{\'e}rez, Fernando and Granger, Brian E and Bussonnier, Matthias and Frederic, Jonathan and Kelley, Kyle and Hamrick, Jessica B and Grout, Jason and Corlay, Sylvain and others},
   booktitle={ELPUB},
+  doi={10.3233/978-1-61499-649-1-87},
   pages={87--90},
   year={2016}
 }
