### 2020-06-25 ihrke/pypillometry#1 @ihrke
sorry, that commit (77f3795) is not related to this issue but to #3. I fixed the commit message but it still shows up here...
