### 2020-06-25 ihrke/pypillometry#3 @ihrke
thanks for spotting those and congrats on the published JOSS paper!
I don't know what the etiquette for the review process is - should I leave it up to you to close the issue, @samhforbes?
