### 2020-06-25 dd9974e
community guidelines (doc)
- added links to issue-tracker to README
- filled out Contributing-section in the documentation

See also: #2
