+++ docs/html/_sources/docs/modeling.ipynb.txt
@@ -4,14 +4,83 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "# Modeling the pupillometric signal"
+    "# Modeling the pupillometric signal\n",
+    "\n",
+    "**WARNING**: This functionality is **experimental**. Some of the provided algorithms are unpublished (or in the process of being published) and may not work well.\n"
    ]
   },
   {
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "TODO"
+    "It can sometimes be useful to think about the pupillometric signal as being composed of different components. One comon assumption, based on the finding that the pupil reflects activity in the norepinephrinergic system, consists of slow, tonic (baseline) and faster, phasic (response) fluctuations. \n"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "import sys\n",
+    "sys.path.insert(0,\"..\")\n",
+    "import pypillometry as pp\n",
+    "import pylab as plt ## access matplotlib functions\n",
+    "\n",
+    "d=pp.PupilData.from_file(\"../data/test.pd\")\n",
+    "faked=pp.create_fake_pupildata(ntrials=20)"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 5,
+   "metadata": {},
+   "outputs": [
+    {
+     "data": {
[…]
