### 2020-06-25 ihrke/pypillometry#2 @ihrke
good points! Can you check whether this commit (dd9974e) addresses this issue sufficiently @samhforbes?
In addition to your requests, I have also added some more information for potential contributors about how the package was designed.
