### 2020-06-25 f7d986a
update installation instructions
- added links to PyStan docs
- added section about using virtual environments

See also: #4
