+++ docs/html/_sources/docs/installation.rst.txt
@@ -1,13 +1,14 @@
 Installation
 ============
 
+Installing :mod:`pypillometry` and its dependencies is automated and can be done by running the following lines (on Mac OS X or Linux). 
+
 .. code-block:: bash
 
-    git clone https://github.com/ihrke/pypillometry.git
-    cd pypillometry
-    pip install -r requirements.txt
-    python setup.py install
-
+    $ git clone https://github.com/ihrke/pypillometry.git
+    $ cd pypillometry
+    $ pip install -r requirements.txt
+    $ python setup.py install
 
 Requirements
 ------------
@@ -21,11 +22,11 @@ It is useful to access :mod:`pypillometry` through Jupyter or Jupyter Notebook, 
 
 All requirements can be installed by running `pip install -r requirements.txt`.
 
-Notes
-^^^^^
+Virtual environments
+--------------------
 
+It can sometimes be useful to install a new package in a new virtual environment using either `Python's virtual environments <https://docs.python.org/3/tutorial/venv.html>`_ or `conda <https://docs.conda.io/en/latest/>`_. 
 
-To install the requirements, either use `pip` as described above, manually install all the packages, or use ``conda``.
 
 .. code-block:: bash
 
@@ -33,15 +34,30 @@ To install the requirements, either use `pip` as described above, manually insta
     $ conda activate pypil
     $ conda install anaconda 
 
-The ``anaconda`` package contains all the requirements except :mod:`pystan`.
-
-To install :mod:`pystan`, I needed to do
+The ``anaconda`` package contains all the requirements except :mod:`pystan` which can be installed from `conda-forge <https://anaconda.org/conda-forge/pystan>`_
 
 .. code-block:: bash
 
-    pip install pystan
-    conda install gcc_linux-64
-    conda install gxx_linux-64
+    $ conda install -c conda-forge pystan
+
+
+Pystan 
+------
+
+Note that the installation of :mod:`pystan` may cause trouble on Windows-systems (you may need to install a compiler). Please follow the instructions on `the Pystan-webpage <https://pystan.readthedocs.io/en/latest/getting_started.html>`_ should you encounter any trouble.
+
+
+Notes/Potential Problems
+-------------------------
+
+Under Linux, I encountered a problem where :mod:`pystan` crashed the `Jupyter kernel <https://jupyter.org/>`_.
+To circumvent this issue, I needed to install :mod:`pystan` using
+
+.. code-block:: bash
+
+    $ pip install pystan
+    $ conda install gcc_linux-64
+    $ conda install gxx_linux-64
 
 otherwise, there were random crashes of the jupyter kernel for some reason. 
 
