### 2020-06-25 ihrke/pypillometry#4 @ihrke
Hei @samhforbes,
can you check whether you are happy with the changes I made to the installation instructions? See f7d986a.
