### 2020-06-25 ihrke/pypillometry#2 @samhforbes
This is great, thanks!
