### 2020-06-25 ihrke/pypillometry#3 @samhforbes
I think you can close issues if you're satisfied, we can always re-open if necessary ;)
This looks good!
