### 2020-06-25 ihrke/pypillometry#4 @samhforbes
@ihrke that's a neat addition - very happy.
