### 2020-06-25 48247b7
bug where npar was interpreted as int instead of float
