+++ pypillometry/pupil.py
@@ -38,6 +38,8 @@ def pupil_kernel_t(t,npar,tmax):
         PRF evaluated at `t`
     """
     t=np.array(t)
+    npar=float(npar)
+    tmax=float(tmax)
     hmax=np.exp(-npar)*tmax**npar ## theoretical maximum
     h = t**(npar) * np.exp(-npar*t / tmax)   #Erlang gamma function Hoek & Levelt (1993)
     h=h/hmax
