### 2020-06-25 b8d21a9
adapted symposium-talk notebook added
this notebook describes the baseline-estimate algorithm

See also: #1
