### 2020-06-25 a86ceff
Update modeling.ipynb
provide an inital documentation of experimental baseline- and response-estimation features

See also: #1
