### 2020-06-25 34e8a3d
updated documentation with experimental modeling features
- added modeling-notebook
- added links to the rest of the documentation to the modeling section

See also: #1
