+++ docs/html/_sources/src/m2r/docs/index.md.txt
@@ -1,0 +1,17 @@
+.. mdinclude:: ../README.md
+
+Contents
+--------
+
+.. toctree::
+    :maxdepth: 1
+
+    example
+    changes
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
