### 2020-06-25 ihrke/pypillometry#1 @ihrke
Hi @samhforbes,

I added some documentation to the modeling-section in the docs and also added a notebook explaining all the details of the experimental baseline- and response-estimation methods.
Let me know if you are happy with this!
