### 2020-06-27 ihrke/pypillometry#5 @szorowi1
Additional installation instructions
Hi @ihrke! First off, this package is great! As a small suggestion, I might consider adding additional installation instructions in the README. For example, folks can install pypillometry directly from Github using:

```bash
pip install git+https://github.com/ihrke/pypillometry.git
```

I was also wondering if you were planning on uploading pypillometry to pypi? This might help make this awesome package more discoverable by python users.
