### 2020-06-27 #2348 @szorowi1
Hi all, had the pleasure of trying out `pypillometry` this week. I was successfully able to install `pypillometry` and then load in my own data, perform preprocessing (blink detection + correction, filtering, rescaling) with ease, and compute/plot ERPDs. The examples were more than enough help to adapt the code to my data, and the API was detailed and easy to read. Overall, the package is very impressive, especially the natural chaining/logging of preprocessing steps. Really cool stuff, @ihrke!

I've made a couple of minor suggestions about installation details here, https://github.com/ihrke/pypillometry/issues/5. I also agree with @samhforbes that more documentation about the modeling functions would be helpful, as is currently being discussed here, https://github.com/ihrke/pypillometry/issues/1.
