### 2020-06-27 ihrke/pypillometry#6 @szorowi1
minor typos
Couple of minor typos in the documentation. Let me know if you'd rather me open a PR.

- [**blinks.ipynb**](https://ihrke.github.io/pypillometry/html/docs/blinks.html):
    - The `API-docs` does not point to a valid URL. Should it be pointed to https://ihrke.github.io/pypillometry/html/docs/api.html?
    - Should the sentence, _"**his** problem can be fixed by setting a stricter value for..."_ start instead with _"**This** problem can be fixed by setting a stricter value..."_?

- [**pipes.ipynb**](https://ihrke.github.io/pypillometry/html/docs/pipes.html):
    - Should _"We see that sampling rate, number **o** datapoints"_ be _"We see that sampling rate, number **of** datapoints"_?

- [**plotting.ipynb**](https://ihrke.github.io/pypillometry/html/docs/plotting.html):
    - Error in recognizing matplotlib, comes up as _":mod:matplotlib-based"_.

- [**summary.ipynb**](https://ihrke.github.io/pypillometry/html/docs/summary.html)
    - Error in recognizing PupilData, comes up as _":class:PupilData"_.
