### 2020-06-27 78c2ca0
Enable PyPI
- created PyPI package
- added installation instructions

See also: #5
