+++ setup.py
@@ -3,9 +3,12 @@ import setuptools
 with open("README.md", "r") as fh:
     long_description = fh.read()
 
+with open("requirements.txt") as fh:
+    requirements=fh.read().split()
+
 setuptools.setup(
     name="pypillometry",
-    version="1.0.1",
+    version="1.0.2",
     author="Matthias Mittner",
     author_email="matthias.mittner@uit.no",
     description="Pupillometry in Python.",
@@ -18,5 +21,6 @@ setuptools.setup(
         "License :: OSI Approved :: MIT License",
         "Operating System :: OS Independent",
     ],
+    install_requires=requirements,
     python_requires='>=3.6',
 )
