### 2020-06-27 ihrke/pypillometry#5 @ihrke
Hi @szorowi1,

very good point, thanks! I have uploaded `pypillometry` to PyPI and it should be possible to install it (and all dependencies using just `pip install pypillometry`).
