### 2020-06-27 7ab1b0a
fix typos
See #6
