### 2020-06-27 ihrke/pypillometry#6 @ihrke
Thanks for spotting those typos! They are corrected in the updated version.
