### 2020-06-27 #2348 @ihrke
Thanks @samhforbes and @szorowi1 for nice comments and suggestions! I have tried to address them in the issues over at ihrke/pypillometry and added the `propose-close` label. If you are happy with the changes, can you close the respective issues?
The remaining issues are

- ihrke/pypillometry#1 (documentation of modeling: I added two notebooks with hopefully helpful information)
- ihrke/pypillometry#5 (installation via pypi: I have added this functionality and updated the installation instructions)
- ihrke/pypillometry#6 (fixed some typos)

Thanks for a great process so far and looking forward to further comments!
