+++ docs/dev.rst
@@ -73,3 +73,25 @@ in the packages' root-directory.
 .. warning::
 
     don't run `make clean`, it will delete the source-files from `docs`
+
+
+Creating Releases
+-----------------
+
+A release should be created whenever crucial bugs have been fixed or new functionality has been added.
+:mod:`pypillometry` goes with version numbers `x.y.z`. Increment `z` for bug-fixes, `y` for new features and
+`x` for releases that break backward-compatibility.
+
+Steps to create a new release:
+
+1. Update/increment the version number in `setup.py` 
+2. create a release on Github: https://github.com/ihrke/pypillometry/releases
+    - this creates a git-tag
+3. create distribution package for `PyPI <https://pypi.org/>`_ and upload package 
+    .. code-block:: bash
+
+        $ python setup.py sdist bdist_wheel
+        $ twine upload dist/pypillometry-x.y.z*
+        
+   
+
