### 2020-06-27 ihrke/pypillometry#6 @szorowi1
Looks good! Feel free to close.
