### 2020-06-28 ihrke/pypillometry#5 @szorowi1
Hi again

On the installation page (https://ihrke.github.io/pypillometry/html/docs/installation.html), I'm not sure if there is supposed to be a line of code following _"pypillometry is on PyPI and released versions can be installed with pip:"_ with the install command. Otherwise, I think it looks great and congrats on the pypi release!
