### 2020-06-29 2b4a9b0
fix installation doc
