### 2020-06-29 ihrke/pypillometry#5 @ihrke
you are right, I messed up some indentation in the `.rst` source-file. I fixed it in 2b4a9b081d2a50f45adb790b1cc58c14020fdc6e
