### 2020-06-29 ihrke/pypillometry#5 @szorowi1
Thanks for the fix. It looks great! Feel free to close the issue if you're happy with it!
