### 2020-06-29 ihrke/pypillometry#1 @samhforbes
I'm satisfied with these changes, it looks interesting and I can get these notebooks to work.
This is in great shape. I'm happy for you to close this, but I'll leave it to you because I don't know whether there are any further thoughts from https://github.com/openjournals/joss-reviews/issues/2348#issuecomment-650473510

Great work!
