### 2020-06-29 ihrke/pypillometry#1 @ihrke
thanks @samhforbes! I'll close this now and we can see whether it needs reopening...
