### 2020-06-30 #2348 @tjmahr
Thank you for this opportunity to review this package. I have written a few lab-specific in-house R packages for managing eyetracking data, so I applaud the effort to provide a general-purpose tool for working with eyetracking data.

I have two disclaimers: I have not formally analyzed pupil data beyond the data-cleaning stage and I seldom use Python.

The documentation is comprehensive and generously provides several notebooks with binder links. The installation instructions helped me use virtual environments for the first time ever. Functions that provide fake data also help learners play with the package.

I have Python 3.6.4 via Anaconda on Windows 10. Installation via cloning from Github and installing dependencies with pip 18.1 went smoothly. I tested the package on some 60 Hz Tobii data I had from an eyetracking task with preschoolers. I created csvs of gaze measurements and event onsets in R. Then in Python I was able to create a plot of pupil diameter over time with trial events annotated. It was cool to move from my own csvs to a nice default plot in a few steps.

I was unable to use velocity based blink detection (it raised an error), so I looked up the API reference for `blink_detect()` and tried
`d.blinks_detect(strategies = ["zero"])`. This worked, but it treated all of the intertrial intervals (during which the eyetracker is not running) in the dataset as blinks (below). Fair enough.

![image](https://user-images.githubusercontent.com/1890315/86173894-52a4d500-bae6-11ea-8384-2794c296e13d.png)

After downsampling to 5 Hz, I noticed that these same blinks were changed to one spurious blink.

![image](https://user-images.githubusercontent.com/1890315/86173990-77994800-bae6-11ea-8ebd-86aa11f80981.png)

At this point, I gave up on trying to use blink detection or interpolation on my data because it was far too noisy and from a task that was never meant to analyze pupil size. That said, the availability of diagnostic plotting, as in `d.blinks_interp_mahot(plot=True, plot_dim=(7,4))`, is a user-friendly addition.

I have to ask: Is there a minimum sampling frequency required for blink analysis? Does the package make any assumptions about the capability of the eyetracker? How should missing eyetracking data be coded in csv file?

The pipeline based processing is a terrific design. By storing the data in a `PupilData` object,  the data can keep track of the changes made to it. And because the history is itself data, it can be applied to other objects with the `apply_history()`. This set of features is awesome for reproducibility and usability.

In eyetracking, we often want to see how gaze measures change in response to some event and adjust times around that event. This package's ERPD (event-related pupil dilation) nicely handles this process by allowing users to select an event and time window before and after this event. It also provides automatic plotting of aggregating gaze measurements. I was not able to get it to work on my noisy Tobii data, but I reproduced the plots in the notebook using the supplied data.

I ran through the phasic and tonic components to confirm that I was able to get pystan to do the modeling.

***

I would recommend this package to any person analyzing pupil data. The history feature is a really smart way to let users prototype a workflow on a single dataset and then quickly apply it to several datasets. My only concern is that it was not clear what requirements are placed on the data.
