### 2020-07-01 ihrke/pypillometry#1 @szorowi1
Just wanted to follow up to say that I've had another look through the modeling docs and I'm satisfied!
