### 2020-07-01 #2348 @szorowi1
Just wanted to say that with the closing of https://github.com/ihrke/pypillometry/issues/1, I'm totally satisfied!

Again, really nice work @ihrke! Will absolutely be recommending this package to the folks collecting pupillometry data in my center
