### 2020-07-01 #2348 @ihrke
Thanks @tjmahr and @szorowi1 for nice and thorough reviews!

Re your experiences with Tobii data, @tjmahr: the question about requirements for data-quality is a good one and one that is hard to answer in general. With the "spotty" data in your figures, many of the blink-features that are based on velocity profiles and detecting points surrounding the blinks are indeed bound to fail. I also see a couple of non-blink artifacts in your dataset which will seriously lead astray the blink-detection/interpolation. Generalizing from these plots to the rest of the data, to make any valid inferences about PD in the presence of so many missing data may be doubtful. One strategy would be to try and interpolate consistent blinks (i.e., with visible pre-post transients) and then go for a more aggressive interpolation-strategy (i.e., interpolate across all missing data) for the remaining time-series.

`pypillometry` was developed using mainly data from three eyetrackers in our lab (two different Eyelinks and a SMI) which is collected during stationary laboratory tasks typically using chin-rests and under constant lighting conditions. Hence, data-quality was likely high and `pypillometry`'s functions will be necessarily tailored and biased towards this data. I hope that with time, as more people and labs start using the package, we will be able to get experience with more diverse datasets and match functionality of the package. Perhaps you would like to upload an example data-file and some details about the problems in an issue over at https://github.com/ihrke/pypillometry? I'll leave the decision about whether this is critical for acceptance of the JOSS-paper or not to you.
