### 2020-07-01 #2348 @tjmahr
Thanks for the response. (I forgot that some eye trackers have chin rests. I was using data from a 4 year old with a free field eye tracker.) I don’t think it’s critical to address for publication.

I’m satisfied. Ship it. 👁
