+++ .github/workflows/python-publish.yml
@@ -22,6 +22,15 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install setuptools wheel twine
+    - name: Update version with latest tag
+      run: git describe --tags > VERSION
+    - name: Commit new VERSION file
+      uses: EndBug/add-and-commit@v4
+      with:
+        message: "update version"
+        add: "VERSION"
+      env:
+        GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
     - name: Build and publish
       env:
         TWINE_USERNAME: ${{ secrets.PYPI_USERNAME }}
